+++ frontend/src/views/DepartmentBatchEnter.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
-import { ref, onMounted } from 'vue'
+import { onMounted, ref } from 'vue'
 import { getDepartment } from '../services/department'
 import { type Attendees, getAttendeesForDepartment } from '../services/attendee'
 import { useToast } from 'vue-toastification'
 import { useRoute } from 'vue-router'
 import type { Department } from '@/services/department'
-import YouthAndAttendeeBatch from '@/components/batch/YouthAndAttendeeBatch.vue'
+import YouthAndAttendeeBatch from '@/components/batch/AttendeeBatchEvent.vue'
 
 const toast = useToast()
 const route = useRoute()
@@ -31,8 +31,11 @@ onMounted(async () => {
       <h1>{{ department?.name }} beitreten</h1>
       <YouthAndAttendeeBatch
         v-if="attendees.youths || attendees.youthLeaders"
-        :youths="attendees.youths"
-        :youthLeader="attendees.youthLeaders"
+        headline="Übernachtungsgäste"
+        :attendeeGroups="[
+          { headline: 'Jugendliche', attendees: attendees.youths },
+          { headline: 'Betreuer', attendees: attendees.youthLeaders }
+        ]"
         :departmentId="departmentId"
         :enterCode="eventCode"
         :leaveCode="leaveCode"
